feat: 폴더 구조 변경 .git/COMMIT_EDITMSG  #79
