--- app/(dashboard)/schedule-dashboard/register/page.tsx
@@ -1,7 +1,0 @@
-import { ScheduleRegisterPage } from "@/src/views/schedule-manage";
-
-const page = () => {
-  return <ScheduleRegisterPage />;
-};
-
-export default page;
