+++ app/(schedule)/schedule-dashboard/manage/page.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { ScheduleManagePage } from "@/src/views/schedule/manage/index";
+
+const page = () => {
+  return <ScheduleManagePage />;
+};
+
+export default page;
