--- src/entities/schedule-card/index.ts
@@ -1,1 +1,0 @@
-export { ShowScheduleCard } from "@/src/entities/schedule-card/ui/show-schedule-card";
