+++ src/entities/schedule-preview-card/api/types.ts
