+++ src/entities/schedule-preview-card/index.ts
@@ -1,0 +1,1 @@
+export { SchedulePreviewCard } from "@/src/entities/schedule-preview-card/ui/schedule-preview-card";
