--- src/features/schedule-manage/create-schedule/index.ts
@@ -1,1 +1,0 @@
-export { CreateSchedule } from "@/src/features/schedule-manage/create-schedule/ui/create-schedule";
