--- src/features/schedule-manage/index.ts
@@ -1,4 +1,0 @@
-export { ShowScheduleCard } from "../../entities/schedule-card/ui/show-schedule-card";
-export { FilterSchedule } from "./ui/filter-schedule";
-export { SelectOption } from "./ui/select";
-export { ScheduleModal } from "./ui/schedule-modal";
