+++ src/features/schedule/manage/schedule-manage-modal/index.ts
@@ -1,0 +1,1 @@
+export { ScheduleManageModal } from "@/src/features/schedule/manage/schedule-manage-modal/ui/schedule-manage-modal";
