+++ src/features/schedule/register/create-schedule/index.ts
@@ -1,0 +1,1 @@
+export { CreateSchedule } from "@/src/features/schedule/register/create-schedule/ui/create-schedule";
