+++ src/features/schedule/register/filter-schedule/index.ts
@@ -1,0 +1,1 @@
+export { FilterSchedule } from "@/src/features/schedule/register/filter-schedule/ui/filter-schedule";
