+++ src/features/schedule/register/filter-schedule/ui/filter-schedule.tsx
