+++ src/features/schedule/register/schedule-calender/index.ts
@@ -1,0 +1,1 @@
+export { ScheduleCalender } from "@/src/features/schedule/register/schedule-calender/ui/schedule-calender";
