+++ src/features/select-option/index.ts
@@ -1,0 +1,1 @@
+export { SelectOption } from "./ui/select-option";
