+++ src/features/select-option/ui/select-option.tsx
