+++ src/views/schedule/register/index.ts
@@ -1,0 +1,1 @@
+export { ScheduleRegisterPage } from "@/src/views/schedule/register/page/schedule-register";
